+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Uliweb Change Log
 * Add `post_do` and `print_sql` function to orm.
 * Change `generic/template_files/common/views.py` to `views.py.tmpl`
 * Add GenericReference value `(Model, object_id)` support
+* Fix pyini save process
 
 0.1.6 Version
 -----------------